fix: 配列を見やすく修正し、ディレクトリ名のタイポを直しました。
