--- app/listeners/RegisteredListener.php
@@ -1,39 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\listeners;
-
-use Illuminate\Auth\Events\Registered;
-use Illuminate\Mail\Mailer;
-use App\Models\User;
-
-class RegisteredListener
-{
-    private $mailer;
-    private $eloquent;
-    /**
-     * Create the event listener.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->mailer = $mailer;
-        $this->eloquent = $eloquent;
-    }
-
-    /**
-     * Handle the event.
-     *
-     * @param  \Illuminate\Auth\Events\Registered  $event
-     * @return void
-     */
-    public function handle(Registered $event)
-    {
-        $user = $this->eloquent->findOrFail($event->user->getAuthIdentifier());
-        $this->mailer->raw('会員登録完了しました', function ($message) use ($user) {
-            $message->subject('会員登録メール')->to($user->email);
-        });
-    }
-}
